requirements/requirements-production.txt
Django==1.10.1
PyYAML==3.12
gunicorn==19.6.0
psycopg2==2.6.2
